+++ crates/incin-macros/src/idx.rs
@@ -38,7 +38,7 @@ pub(crate) fn idx(input: TokenStream) -> TokenStream {
                             .into();
                         }
                     };
-                    let path = quote! { incin::prelude:: };
+                    let path = quote! { ::incin::prelude:: };
                     crate::shape::lit_to_typenum(val, &path)
                 } else {
                     return syn::Error::new_spanned(
@@ -65,7 +65,7 @@ pub(crate) fn idx(input: TokenStream) -> TokenStream {
                                 }
                             };
                             if val == 1 {
-                                quote! { incin::prelude::InferDim }
+                                quote! { ::incin::prelude::InferDim }
                             } else {
                                 return syn::Error::new_spanned(
                                     lit_int,
@@ -92,7 +92,7 @@ pub(crate) fn idx(input: TokenStream) -> TokenStream {
             }
             Expr::Path(expr_path) => {
                 if let Some(ident) = expr_path.path.get_ident() {
-                    quote! { incin::prelude::NamedDyn<#ident> }
+                    quote! { ::incin::prelude::NamedDyn<#ident> }
                 } else {
                     return syn::Error::new_spanned(
                         &expr_path,
@@ -104,7 +104,7 @@ pub(crate) fn idx(input: TokenStream) -> TokenStream {
             }
             Expr::Range(expr_range) => match (&expr_range.start, &expr_range.end) {
                 (None, None) => {
-                    quote! { incin::prelude::Ellipsis }
+                    quote! { ::incin::prelude::Ellipsis }
                 }
                 (Some(start), Some(end)) => {
                     let start_val = if let Expr::Lit(expr_lit) = &**start {
@@ -175,11 +175,11 @@ pub(crate) fn idx(input: TokenStream) -> TokenStream {
                     }
 
                     let diff = end_val - start_val;
-                    let path = quote! { incin::prelude:: };
+                    let path = quote! { ::incin::prelude:: };
                     let start_type = crate::shape::lit_to_typenum(start_val, &path);
                     let end_type = crate::shape::lit_to_typenum(end_val, &path);
                     let diff_type = crate::shape::lit_to_typenum(diff, &path);
-                    quote! { incin::prelude::Slice<#start_type, #end_type, #diff_type> }
+                    quote! { ::incin::prelude::Slice<#start_type, #end_type, #diff_type> }
                 }
                 _ => {
                     return syn::Error::new_spanned(